+++ src/libserveez/hash.h
@@ -35,12 +35,12 @@ typedef struct svz_hash svz_hash_t;
  */
 struct svz_hash
 {
-  int buckets;                     /* number of buckets in the table */
-  int fill;                        /* number of filled buckets */
-  int keys;                        /* number of stored keys */
+  size_t buckets;                  /* number of buckets in the table */
+  size_t fill;                     /* number of filled buckets */
+  size_t keys;                     /* number of stored keys */
   int (* equals) (char *, char *); /* key string equality callback */
   unsigned long (* code) (char *); /* hash code calculation callback */
-  unsigned (* keylen) (char *);    /* how to get the hash key length */
+  size_t (* keylen) (char *);      /* how to get the hash key length */
   svz_free_func_t destroy;         /* element destruction callback */
   svz_hash_bucket_t *table;        /* hash table */
 };
@@ -53,13 +53,13 @@ __BEGIN_DECLS
 /*
  * Basic hash table functions.
  */
-SERVEEZ_API svz_hash_t *svz_hash_create (int, svz_free_func_t);
+SERVEEZ_API svz_hash_t *svz_hash_create (size_t, svz_free_func_t);
 SERVEEZ_API void svz_hash_destroy (svz_hash_t *);
 SERVEEZ_API void *svz_hash_delete (svz_hash_t *, char *);
 SERVEEZ_API void *svz_hash_put (svz_hash_t *, char *, void *);
 SERVEEZ_API void *svz_hash_get (const svz_hash_t *, char *);
 SERVEEZ_API void svz_hash_foreach (svz_hash_do_t *, svz_hash_t *, void *);
-SERVEEZ_API int svz_hash_size (const svz_hash_t *);
+SERVEEZ_API size_t svz_hash_size (const svz_hash_t *);
 SERVEEZ_API char *svz_hash_contains (const svz_hash_t *, void *);
 SERVEEZ_API int svz_hash_exists (const svz_hash_t *, char *);
